create the navigation folder
